Turned it in without making adequate swim lane diagrams
To avoid being late I had to stop working on the project when it still was in a potentially less-than-A state.

The most glaringly lacking section was part H-H4 Swimming Lane Diagrams. I only made one and it captured the login process. It was also hard to see (if possible at all).